pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "bcrypt>=4.3.0",
    "flask>=3.1.0",
    "flask-cors>=5.0.1",
    "flask-socketio>=5.5.1",
    "openai>=1.76.0",
    "pyjwt>=2.10.1",
    "pymongo>=4.12.0",
    "python-dotenv>=1.1.0",
]
